requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymongo
pydantic[email]
